Разбивает main.js на модули и подключает их.
