+++ src/components/cost.js
@@ -1,0 +1,7 @@
+export const createTripInfoCostElement = () => {
+  return (
+    `<p class="trip-info__cost">
+      Total: &euro;&nbsp;<span class="trip-info__cost-value">1230</span>
+    </p>`
+  );
+};
