+++ src/components/info.js
@@ -1,0 +1,16 @@
+const createTripMainInfoSectionElement = () => {
+  return `<section class="trip-main__trip-info  trip-info"></section>`;
+};
+
+// Информация о маршруте (доп.задание);
+const createTripMainInfoElement = () => {
+  return (
+    `<div class="trip-info__main">
+      <h1 class="trip-info__title">Amsterdam &mdash; Chamonix &mdash; Geneva</h1>
+
+      <p class="trip-info__dates">Mar 18&nbsp;&mdash;&nbsp;20</p>
+    </div>`
+  );
+};
+
+export {createTripMainInfoSectionElement, createTripMainInfoElement};
